pyproject.toml
[tool.poetry]
name = "djangodrf-homework"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.28.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
